requirements.txt
apispec==5.2.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
Flask==2.1.2
flask-smorest==0.38.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.16.0
MpesaRest==0.0.2
packaging==21.3
pika==1.2.1
prometheus-client==0.14.1
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.9
requests==2.28.0
SQLAlchemy==1.4.37
urllib3==1.26.9
webargs==8.1.0
Werkzeug==2.1.2
zipp==3.8.0
